refactor: Move ClientShopPage logic to ShopPage and enhance animations
